+++ KeyboardKitDemoKeyboard/Buttons/DemoButton.swift
@@ -1,6 +1,6 @@
 //
 //  DemoButton.swift
-//  KeyboardKitExampleKeyboard
+//  KeyboardKitDemoKeyboard
 //
 //  Created by Daniel Saidi on 2019-04-30.
 //  Copyright © 2019 Daniel Saidi. All rights reserved.