+++ KeyboardKitDemoKeyboard/DemoKeyboardActionHandler.swift
@@ -1,6 +1,6 @@
 //
 //  DemoKeyboardActionHandler.swift
-//  KeyboardKitExampleKeyboard
+//  KeyboardKitDemoKeyboard
 //
 //  Created by Daniel Saidi on 2019-04-24.
 //  Copyright © 2019 Daniel Saidi. All rights reserved.