+++ KeyboardKitDemoKeyboard/Keyboards/AlphabeticKeyboard.swift
@@ -1,6 +1,6 @@
 //
 //  AlphabeticKeyboard.swift
-//  KeyboardKitExampleKeyboard
+//  KeyboardKitDemoKeyboard
 //
 //  Created by Daniel Saidi on 2019-05-13.
 //  Copyright © 2019 Daniel Saidi. All rights reserved.