+++ KeyboardKitDemoKeyboard_SwiftUI/DemoKeyboardActionHandler.swift
@@ -1,6 +1,6 @@
 //
 //  DemoKeyboardActionHandler.swift
-//  KeyboardKitExampleKeyboard_SwiftUI
+//  KeyboardKitDemoKeyboard_SwiftUI
 //
 //  Created by Daniel Saidi on 2020-07-02.
 //  Copyright © 2020 Daniel Saidi. All rights reserved.