+++ Sources/KeyboardKit/Autocomplete/AutocompleteSuggestions.swift
@@ -1,0 +1,14 @@
+//
+//  AutocompleteSuggestions.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2020-09-12.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This typealias is used to handle autocomplete suggestions.
+ */
+public typealias AutocompleteSuggestions = [String]